Refactor code structure: extract styles, shared modal, move utils outside app/
- Extract all StyleSheet.create blocks to co-located .styles.js files for all screens
- Create shared SessionDetailsModal component in components/sessions/ to eliminate
  duplicate modal code between SessionsListScreen and MySessionsScreen
- Move dateTimeUtils.js and sessionUtils.js from app/utils/ to mobile/utils/ to fix
  Expo Router warning about missing default exports (Expo treats app/ files as routes)
- Remove debug console.log statements from api.js interceptors and server.js
- Move dev scripts (add-test-users.js, add-more-users.js) to backend/scripts/
- Add show/hide password toggle to LoginScreen
- Fix section title colors on ProfileScreen (now visible)
- Remove logout button from MySessionsScreen; fix logout redirect to login screen
- Add safe area insets to modals on SessionsListScreen, MySessionsScreen, ProfileScreen
- Add Cancel/Leave/Invite Friends buttons to MySessionsScreen session modal
