+++ backend/scripts/README.md
@@ -1,0 +1,2 @@
+# Dev Scripts
+One-time development scripts. Do not run in production.
