+++ mobile/app/screens/auth/LoginScreen.js
@@ -9,11 +9,12 @@ import {
   ScrollView,
   ImageBackground,
 } from 'react-native';
+import styles from './LoginScreen.styles';
 import { useRouter } from 'expo-router';
 import { AuthContext } from '../../../contexts/AuthContext';
 import { useTheme } from '../../../contexts/ThemeContext';
 
-const LoginScreen = ({ navigation = null, showSignup = false }) => {
+const LoginScreen = ({ showSignup = false }) => {
   const { login, signup } = useContext(AuthContext);
   const theme = useTheme();
   const router = useRouter();
@@ -165,115 +166,5 @@ const LoginScreen = ({ navigation = null, showSignup = false }) => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  backgroundImage: {
-    flex: 1,
-  },
-  overlay: {
-    ...StyleSheet.absoluteFillObject,
-    backgroundColor: 'rgba(0, 0, 0, 0.4)',
-  },
-  container: {
-    flex: 1,
-  },
-  contentContainer: {
-    flexGrow: 1,
-    justifyContent: 'center',
-    padding: 20,
-  },
-  header: {
-    marginBottom: 40,
-    alignItems: 'center',
-  },
-  title: {
-    fontSize: 32,
-    fontWeight: 'bold',
-    color: '#2196F3',
-  },
-  subtitle: {
-    fontSize: 14,
-    color: '#666',
-    marginTop: 8,
-  },
-  form: {
-    backgroundColor: 'white',
-    borderRadius: 12,
-    padding: 20,
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.1,
-    shadowRadius: 3,
-    elevation: 3,
-  },
-  label: {
-    fontSize: 14,
-    fontWeight: '600',
-    color: '#333',
-    marginBottom: 8,
-  },
-  input: {
-    borderWidth: 2,
-    borderColor: '#ddd',
-    borderRadius: 8,
-    padding: 12,
-    marginBottom: 16,
-    fontSize: 14,
-    backgroundColor: 'white',
-    color: '#333',
-  },
-  passwordContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    borderWidth: 2,
-    borderColor: '#ddd',
-    borderRadius: 8,
-    backgroundColor: 'white',
-    marginBottom: 16,
-  },
-  passwordInput: {
-    flex: 2,
-    padding: 12,
-    fontSize: 14,
-    color: '#333',
-  },
-  eyeButton: {
-    paddingHorizontal: 12,
-    paddingVertical: 12,
-  },
-  eyeIcon: {
-    fontSize: 18,
-  },
-  button: {
-    backgroundColor: '#2196F3',
-    borderRadius: 8,
-    padding: 14,
-    alignItems: 'center',
-    marginTop: 8,
-  },
-  buttonDisabled: {
-    opacity: 0.6,
-  },
-  buttonText: {
-    color: 'white',
-    fontSize: 16,
-    fontWeight: '600',
-  },
-  link: {
-    color: '#2196F3',
-    textAlign: 'center',
-    marginTop: 16,
-    fontSize: 14,
-  },
-  forgotPasswordLink: {
-    alignSelf: 'flex-end',
-    marginTop: -8,
-    marginBottom: 16,
-  },
-  forgotPasswordText: {
-    color: '#2196F3',
-    fontSize: 14,
-  },
-});
 
 export default LoginScreen;
