+++ mobile/app/screens/sessions/CreateSessionScreen.js
@@ -12,11 +12,12 @@ import {
   FlatList,
   Platform,
 } from 'react-native';
+import styles from './CreateSessionScreen.styles';
 import DateTimePicker from '@react-native-community/datetimepicker';
 import { AuthContext } from '../../../contexts/AuthContext';
 import { useTheme } from '../../../contexts/ThemeContext';
 import { sessionAPI, venueAPI } from '../../../services/api';
-import { checkUserCreationConflict } from '../../utils/sessionUtils';
+import { checkUserCreationConflict } from '../../../utils/sessionUtils';
 
 const CreateSessionScreen = ({ navigation }) => {
   const theme = useTheme();
@@ -508,187 +509,5 @@ const CreateSessionScreen = ({ navigation }) => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  backgroundImage: {
-    flex: 1,
-  },
-  overlay: {
-    ...StyleSheet.absoluteFillObject,
-    backgroundColor: 'rgba(0, 0, 0, 0.4)',
-  },
-  container: {
-    flex: 1,
-  },
-  contentContainer: {
-    flexGrow: 1,
-  },
-  header: {
-    backgroundColor: 'rgba(33, 150, 243, 0.6)',
-    paddingTop: 50,
-    paddingHorizontal: 16,
-    paddingBottom: 16,
-  },
-  backButtonBox: {
-    borderWidth: 2,
-    borderRadius: 8,
-    paddingHorizontal: 12,
-    paddingVertical: 8,
-    marginBottom: 12,
-    alignSelf: 'flex-start',
-  },
-  backButton: {
-    fontSize: 16,
-    fontWeight: '600',
-  },
-  title: {
-    fontSize: 28,
-    fontWeight: 'bold',
-    color: 'white',
-  },
-  form: {
-    padding: 16,
-  },
-  label: {
-    fontSize: 14,
-    fontWeight: '600',
-    color: '#333',
-    marginBottom: 8,
-  },
-  input: {
-    borderWidth: 1,
-    borderRadius: 8,
-    padding: 12,
-    marginBottom: 16,
-    fontSize: 14,
-  },
-  textArea: {
-    height: 80,
-    textAlignVertical: 'top',
-  },
-  row: {
-    flexDirection: 'row',
-    gap: 12,
-  },
-  halfInput: {
-    flex: 1,
-  },
-  hint: {
-    fontSize: 12,
-    color: '#999',
-    marginTop: -8,
-    marginBottom: 16,
-  },
-  actions: {
-    gap: 12,
-    marginTop: 16,
-  },
-  button: {
-    paddingVertical: 14,
-    borderRadius: 8,
-    alignItems: 'center',
-  },
-  createButton: {
-    backgroundColor: '#4CAF50',
-  },
-  cancelButton: {
-    backgroundColor: '#e0e0e0',
-  },
-  buttonText: {
-    color: 'white',
-    fontWeight: '600',
-    fontSize: 16,
-  },
-  cancelButtonText: {
-    color: '#666',
-    fontWeight: '600',
-    fontSize: 16,
-  },
-  buttonDisabled: {
-    opacity: 0.6,
-  },
-  dropdown: {
-    borderWidth: 1,
-    borderRadius: 8,
-    padding: 12,
-    marginBottom: 16,
-    fontSize: 14,
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-  },
-  dropdownText: {
-    fontSize: 14,
-    flex: 1,
-  },
-  dropdownArrow: {
-    fontSize: 14,
-    marginLeft: 8,
-  },
-  modalOverlay: {
-    flex: 1,
-    justifyContent: 'flex-end',
-  },
-  modalContent: {
-    borderTopLeftRadius: 16,
-    borderTopRightRadius: 16,
-    paddingTop: 16,
-    maxHeight: '80%',
-  },
-  modalHeader: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    paddingHorizontal: 16,
-    paddingBottom: 12,
-    borderBottomWidth: 1,
-    borderBottomColor: '#eee',
-  },
-  modalTitle: {
-    fontSize: 18,
-    fontWeight: 'bold',
-  },
-  modalClose: {
-    fontSize: 24,
-    fontWeight: 'bold',
-  },
-  modalOption: {
-    borderBottomWidth: 1,
-    padding: 16,
-  },
-  modalOptionSelected: {
-    backgroundColor: 'rgba(33, 150, 243, 0.1)',
-  },
-  modalOptionText: {
-    fontSize: 14,
-  },
-  modalSubtext: {
-    fontSize: 12,
-    marginTop: 4,
-  },
-  datePickerButtons: {
-    flexDirection: 'row',
-    justifyContent: 'flex-end',
-    paddingHorizontal: 16,
-    paddingVertical: 8,
-  },
-  datePickerButton: {
-    color: '#2196F3',
-    fontSize: 16,
-    fontWeight: '600',
-    paddingHorizontal: 16,
-    paddingVertical: 8,
-  },
-  timeOptionContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-  },
-  nextDayLabel: {
-    fontSize: 12,
-    fontStyle: 'italic',
-    marginLeft: 8,
-  },
-});
 
 export default CreateSessionScreen;
