+++ mobile/app/screens/sessions/SessionDetailScreen.js
@@ -2,15 +2,15 @@ import React, { useState, useEffect, useContext } from 'react';
 import {
   View,
   Text,
-  StyleSheet,
   TouchableOpacity,
   Alert,
   ActivityIndicator,
   ScrollView,
 } from 'react-native';
+import styles from './SessionDetailScreen.styles';
 import { AuthContext } from '../../../contexts/AuthContext';
 import { sessionAPI } from '../../../services/api';
-import { checkUserTimeConflict } from '../../utils/sessionUtils';
+import { checkUserTimeConflict } from '../../../utils/sessionUtils';
 
 const SessionDetailScreen = ({ route, navigation }) => {
   const { sessionId } = route.params;
@@ -188,141 +188,5 @@ const SessionDetailScreen = ({ route, navigation }) => {
     </ScrollView>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#f5f5f5',
-  },
-  header: {
-    backgroundColor: '#2196F3',
-    paddingTop: 50,
-    paddingHorizontal: 16,
-    paddingBottom: 16,
-  },
-  backButton: {
-    color: 'white',
-    fontSize: 16,
-    marginBottom: 12,
-  },
-  title: {
-    fontSize: 28,
-    fontWeight: 'bold',
-    color: 'white',
-  },
-  content: {
-    padding: 16,
-  },
-  card: {
-    backgroundColor: 'white',
-    borderRadius: 12,
-    padding: 16,
-    marginBottom: 12,
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.1,
-    shadowRadius: 3,
-    elevation: 3,
-  },
-  cardTitle: {
-    fontSize: 16,
-    fontWeight: 'bold',
-    color: '#333',
-    marginBottom: 12,
-  },
-  row: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    marginBottom: 12,
-    paddingBottom: 12,
-    borderBottomWidth: 1,
-    borderBottomColor: '#eee',
-  },
-  label: {
-    fontSize: 14,
-    color: '#666',
-  },
-  value: {
-    fontSize: 14,
-    fontWeight: '600',
-    color: '#333',
-    flex: 1,
-    textAlign: 'right',
-  },
-  costCard: {
-    backgroundColor: '#e3f2fd',
-    borderRadius: 12,
-    padding: 16,
-    marginBottom: 12,
-  },
-  costLabel: {
-    fontSize: 12,
-    color: '#666',
-    marginBottom: 4,
-  },
-  costValue: {
-    fontSize: 24,
-    fontWeight: 'bold',
-    color: '#2196F3',
-    marginBottom: 16,
-  },
-  emptyText: {
-    fontSize: 14,
-    color: '#999',
-  },
-  participant: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    paddingVertical: 12,
-    borderBottomWidth: 1,
-    borderBottomColor: '#eee',
-  },
-  participantName: {
-    fontSize: 14,
-    fontWeight: '600',
-    color: '#333',
-  },
-  participantEmail: {
-    fontSize: 12,
-    color: '#999',
-    marginTop: 4,
-  },
-  participantCost: {
-    fontSize: 14,
-    fontWeight: '600',
-    color: '#2196F3',
-  },
-  actionContainer: {
-    padding: 16,
-  },
-  button: {
-    paddingVertical: 14,
-    borderRadius: 8,
-    alignItems: 'center',
-  },
-  joinButton: {
-    backgroundColor: '#4CAF50',
-  },
-  leaveButton: {
-    backgroundColor: '#FF9800',
-  },
-  cancelButton: {
-    backgroundColor: '#f44336',
-  },
-  buttonText: {
-    color: 'white',
-    fontWeight: '600',
-    fontSize: 16,
-  },
-  buttonDisabled: {
-    opacity: 0.6,
-  },
-  errorText: {
-    fontSize: 16,
-    color: '#f44336',
-    textAlign: 'center',
-  },
-});
 
 export default SessionDetailScreen;
