+++ mobile/app/screens/sessions/SessionsListScreen.styles.js
@@ -1,0 +1,479 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+  },
+  overlay: {
+    ...StyleSheet.absoluteFillObject,
+    backgroundColor: 'rgba(0, 0, 0, 0.3)',
+  },
+  header: {
+    backgroundColor: 'rgba(33, 150, 243, 0.6)',
+    paddingTop: 50,
+    paddingHorizontal: 16,
+    paddingBottom: 20,
+  },
+  headerTitle: {
+    fontSize: 28,
+    fontWeight: 'bold',
+    color: 'white',
+  },
+  headerSubtitle: {
+    fontSize: 14,
+    color: 'rgba(255,255,255,0.7)',
+    marginTop: 4,
+  },
+  headerButtons: {
+    flexDirection: 'row',
+    marginTop: 12,
+    gap: 8,
+  },
+  createButton: {
+    backgroundColor: '#4CAF50',
+    paddingHorizontal: 12,
+    paddingVertical: 8,
+    borderRadius: 6,
+    flex: 1,
+  },
+  createButtonText: {
+    color: 'white',
+    fontWeight: '600',
+    textAlign: 'center',
+  },
+  logoutButton: {
+    backgroundColor: '#f44336',
+    paddingHorizontal: 12,
+    paddingVertical: 8,
+    borderRadius: 6,
+    flex: 1,
+  },
+  logoutButtonText: {
+    color: 'white',
+    fontWeight: '600',
+    textAlign: 'center',
+  },
+  filterContainer: {
+    flexDirection: 'row',
+    paddingHorizontal: 16,
+    paddingTop: 12,
+    paddingBottom: 8,
+    gap: 8,
+  },
+  filterButton: {
+    paddingHorizontal: 12,
+    paddingVertical: 8,
+    borderRadius: 6,
+    backgroundColor: 'white',
+    borderWidth: 1,
+    borderColor: '#ddd',
+    flex: 1,
+  },
+  filterButtonActive: {
+    backgroundColor: '#2196F3',
+    borderColor: '#2196F3',
+  },
+  filterText: {
+    textAlign: 'center',
+    color: '#666',
+    fontWeight: '500',
+  },
+  filterTextActive: {
+    textAlign: 'center',
+    color: 'white',
+    fontWeight: '500',
+  },
+  dateFilterContainer: {
+    flexDirection: 'row',
+    paddingHorizontal: 16,
+    paddingBottom: 12,
+    alignItems: 'center',
+    gap: 8,
+  },
+  datePickerButton: {
+    flex: 1,
+    flexDirection: 'row',
+    alignItems: 'center',
+    backgroundColor: 'white',
+    borderRadius: 8,
+    paddingHorizontal: 16,
+    paddingVertical: 12,
+    borderWidth: 1,
+    borderColor: '#2196F3',
+    gap: 10,
+  },
+  datePickerIcon: {
+    fontSize: 20,
+  },
+  datePickerText: {
+    fontSize: 15,
+    color: '#333',
+    fontWeight: '500',
+  },
+  clearDateButton: {
+    backgroundColor: '#f44336',
+    paddingHorizontal: 14,
+    paddingVertical: 12,
+    borderRadius: 8,
+  },
+  clearDateText: {
+    color: 'white',
+    fontWeight: '600',
+    fontSize: 16,
+  },
+  datePickerModalOverlay: {
+    flex: 1,
+    backgroundColor: 'rgba(0, 0, 0, 0.5)',
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  datePickerModalContent: {
+    backgroundColor: 'white',
+    borderRadius: 16,
+    padding: 20,
+    width: '90%',
+    maxWidth: 400,
+  },
+  datePickerHeader: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    marginBottom: 15,
+  },
+  datePickerTitle: {
+    fontSize: 18,
+    fontWeight: 'bold',
+    color: '#333',
+  },
+  datePickerDone: {
+    fontSize: 16,
+    color: '#2196F3',
+    fontWeight: '600',
+  },
+  datePicker: {
+    width: '100%',
+  },
+  locationFilterContainer: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    paddingHorizontal: 16,
+    paddingBottom: 12,
+    gap: 8,
+  },
+  locationDropdownButton: {
+    flex: 1,
+    flexDirection: 'row',
+    alignItems: 'center',
+    backgroundColor: 'white',
+    borderRadius: 8,
+    paddingHorizontal: 12,
+    paddingVertical: 12,
+    borderWidth: 1,
+    borderColor: '#2196F3',
+    gap: 8,
+  },
+  locationDropdownButtonActive: {
+    backgroundColor: '#E3F2FD',
+    borderColor: '#1565C0',
+  },
+  locationFilterIcon: {
+    fontSize: 16,
+  },
+  locationDropdownText: {
+    flex: 1,
+    fontSize: 14,
+    color: '#666',
+    fontWeight: '500',
+  },
+  locationDropdownTextActive: {
+    color: '#1565C0',
+    fontWeight: '600',
+  },
+  locationDropdownChevron: {
+    fontSize: 14,
+    color: '#999',
+  },
+  clearLocationButton: {
+    backgroundColor: '#f44336',
+    paddingHorizontal: 14,
+    paddingVertical: 12,
+    borderRadius: 8,
+  },
+  locationModalOverlay: {
+    flex: 1,
+    backgroundColor: 'rgba(0,0,0,0.5)',
+    justifyContent: 'center',
+    alignItems: 'center',
+    padding: 24,
+  },
+  locationModalContent: {
+    backgroundColor: 'white',
+    borderRadius: 16,
+    padding: 8,
+    width: '100%',
+    maxHeight: '70%',
+  },
+  locationModalTitle: {
+    fontSize: 16,
+    fontWeight: 'bold',
+    color: '#333',
+    paddingHorizontal: 16,
+    paddingVertical: 14,
+    borderBottomWidth: 1,
+    borderBottomColor: '#eee',
+  },
+  locationOption: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    paddingHorizontal: 16,
+    paddingVertical: 14,
+    borderBottomWidth: 1,
+    borderBottomColor: '#f5f5f5',
+  },
+  locationOptionSelected: {
+    backgroundColor: '#E3F2FD',
+  },
+  locationOptionContent: {
+    flex: 1,
+  },
+  locationOptionText: {
+    fontSize: 15,
+    color: '#333',
+    fontWeight: '500',
+  },
+  locationOptionTextSelected: {
+    color: '#1565C0',
+    fontWeight: '700',
+  },
+  locationOptionAddress: {
+    fontSize: 12,
+    color: '#999',
+    marginTop: 2,
+  },
+  locationOptionCheck: {
+    fontSize: 16,
+    color: '#1565C0',
+    fontWeight: 'bold',
+    marginLeft: 8,
+  },
+  datePickerActions: {
+    marginTop: 15,
+    alignItems: 'center',
+  },
+  datePickerClearButton: {
+    paddingVertical: 10,
+    paddingHorizontal: 20,
+    backgroundColor: '#f44336',
+    borderRadius: 8,
+  },
+  datePickerClearText: {
+    color: 'white',
+    fontWeight: '600',
+    fontSize: 14,
+  },
+  listContainer: {
+    flex: 1,
+  },
+  listContent: {
+    paddingHorizontal: 16,
+    paddingTop: 8,
+    paddingBottom: 16,
+  },
+  sessionCard: {
+    backgroundColor: 'white',
+    borderRadius: 12,
+    padding: 16,
+    marginBottom: 12,
+    shadowColor: '#000',
+    shadowOffset: { width: 0, height: 2 },
+    shadowOpacity: 0.1,
+    shadowRadius: 3,
+    elevation: 3,
+  },
+  sessionHeader: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    marginBottom: 8,
+  },
+  sessionTitle: {
+    fontSize: 16,
+    fontWeight: 'bold',
+    color: '#333',
+    flex: 1,
+  },
+  status: {
+    paddingHorizontal: 8,
+    paddingVertical: 4,
+    borderRadius: 4,
+    fontSize: 11,
+    fontWeight: '600',
+  },
+  openStatus: {
+    backgroundColor: '#c8e6c9',
+    color: '#2e7d32',
+  },
+  fullStatus: {
+    backgroundColor: '#ffcccc',
+    color: '#c62828',
+  },
+  location: {
+    fontSize: 13,
+    color: '#666',
+    marginBottom: 12,
+  },
+  sessionDetails: {
+    flexDirection: 'row',
+    justifyContent: 'space-around',
+    paddingTop: 12,
+    borderTopWidth: 1,
+    borderTopColor: '#eee',
+  },
+  detailLabel: {
+    fontSize: 11,
+    color: '#999',
+    marginBottom: 4,
+  },
+  detailValue: {
+    fontSize: 14,
+    fontWeight: '600',
+    color: '#333',
+  },
+  buttonContainer: {
+    flexDirection: 'row',
+    gap: 8,
+    marginTop: 12,
+  },
+  actionButton: {
+    flex: 1,
+    paddingVertical: 12,
+    paddingHorizontal: 8,
+    borderRadius: 8,
+    alignItems: 'center',
+  },
+  joinButton: {
+    backgroundColor: '#4CAF50',
+  },
+  leaveButton: {
+    backgroundColor: '#f44336',
+  },
+  cancelButton: {
+    backgroundColor: '#FF9800',
+  },
+  buttonDisabled: {
+    opacity: 0.4,
+  },
+  actionButtonText: {
+    color: 'white',
+    fontWeight: '600',
+    fontSize: 13,
+  },
+  loaderContainer: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  emptyContainer: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    paddingHorizontal: 32,
+  },
+  emptyText: {
+    fontSize: 16,
+    color: '#999',
+    marginBottom: 16,
+  },
+  emptyButton: {
+    backgroundColor: '#2196F3',
+    paddingHorizontal: 24,
+    paddingVertical: 12,
+    borderRadius: 6,
+  },
+  emptyButtonText: {
+    color: 'white',
+    fontWeight: '600',
+  },
+  codeButton: {
+    backgroundColor: '#9C27B0',
+    paddingHorizontal: 12,
+    paddingVertical: 8,
+    borderRadius: 6,
+    flex: 1,
+  },
+  codeButtonText: {
+    color: 'white',
+    fontWeight: '600',
+    textAlign: 'center',
+    fontSize: 13,
+  },
+  codeModalOverlay: {
+    flex: 1,
+    backgroundColor: 'rgba(0, 0, 0, 0.6)',
+    justifyContent: 'center',
+    alignItems: 'center',
+    padding: 24,
+  },
+  codeModalContent: {
+    backgroundColor: 'white',
+    borderRadius: 16,
+    padding: 24,
+    width: '100%',
+  },
+  codeModalTitle: {
+    fontSize: 22,
+    fontWeight: 'bold',
+    color: '#333',
+    marginBottom: 8,
+    textAlign: 'center',
+  },
+  codeModalSubtitle: {
+    fontSize: 14,
+    color: '#666',
+    marginBottom: 20,
+    textAlign: 'center',
+  },
+  codeInput: {
+    borderWidth: 1.5,
+    borderColor: '#9C27B0',
+    borderRadius: 10,
+    paddingVertical: 14,
+    paddingHorizontal: 16,
+    fontSize: 18,
+    letterSpacing: 1,
+    color: '#333',
+    marginBottom: 20,
+    textAlign: 'center',
+  },
+  codeModalButtons: {
+    flexDirection: 'row',
+    gap: 12,
+  },
+  codeModalCancel: {
+    flex: 1,
+    paddingVertical: 14,
+    borderRadius: 10,
+    borderWidth: 1.5,
+    borderColor: '#ddd',
+    alignItems: 'center',
+  },
+  codeModalCancelText: {
+    color: '#666',
+    fontWeight: '600',
+    fontSize: 15,
+  },
+  codeModalConfirm: {
+    flex: 1,
+    paddingVertical: 14,
+    borderRadius: 10,
+    backgroundColor: '#9C27B0',
+    alignItems: 'center',
+  },
+  codeModalConfirmText: {
+    color: 'white',
+    fontWeight: '600',
+    fontSize: 15,
+  },
+});
+
+export default styles;
