+++ mobile/services/api.js
@@ -19,7 +19,6 @@ api.interceptors.request.use(
     } catch (err) {
       console.warn('AsyncStorage error:', err);
     }
-    console.log('API Request:', config.url, config.method);
     return config;
   },
   (error) => {
@@ -31,7 +30,6 @@ api.interceptors.request.use(
 // Add response error logging
 api.interceptors.response.use(
   (response) => {
-    console.log('API Response:', response.status, response.data);
     return response;
   },
   (error) => {
